fix: further tidy test organisation
